chore: move and renme check and radio groups
