+++ packages/react/src/components/form/Checkbox/Checkbox.mdx
@@ -4,7 +4,7 @@ import { Do, Dont, Stack } from '@doc-components';
 import * as TableStories from '../../Table/Table.stories';
 
 import * as CheckboxStories from './Checkbox.stories';
-import * as CheckboxGroupStories from './Group/Group.stories';
+import * as CheckboxGroupStories from './CheckboxGroup.stories';
 
 <Meta of={CheckboxStories} />
 
@@ -165,3 +165,4 @@ Det skal være mulig å flytte fokus til `Checkbox`, ved å trykke `Tab` på tas
 `Checkbox` skal alltid ha en `label`. Bruker du `label` fra en annen plass, husk da å bruke `aria-label` eller `aria-labelledby` (Se kode under).
 
 <Canvas of={CheckboxStories.Single} />
+./Group.stories
