--- packages/react/src/components/form/Checkbox/Group/index.ts
@@ -1,1 +1,0 @@
-export * from './Group';
