--- packages/react/src/components/form/Radio/Group/index.ts
@@ -1,1 +1,0 @@
-export * from './Group';
