+++ packages/react/src/components/form/Radio/Radio.mdx
@@ -1,7 +1,7 @@
 import { Meta, Canvas, Controls, Primary } from '@storybook/blocks';
 
 import * as RadioStories from './Radio.stories';
-import * as RadioGroupStories from './Group/Group.stories';
+import * as RadioGroupStories from './RadioGroup.stories';
 
 <Meta of={RadioStories} />
 
@@ -122,3 +122,4 @@ Merk at vi bruker piltastene til å velge alternativer i `Radio`-komponenten, ik
 - Pil ned/høyre - går til neste valg
 - Mellomromstasten - går til det første alternativet i en gruppe med alternativknapper, når gruppa ikke har et forhåndsvalgt alternativ.
 - Tab - går til hele gruppa med alternativer.
+./Group.stories
