+++ packages/react/src/components/form/Radio/RadioGroup.tsx
@@ -2,8 +2,8 @@ import cl from 'clsx/lite';
 import type { ReactNode } from 'react';
 import { createContext, forwardRef, useId } from 'react';
 
-import type { FieldsetProps } from '../../Fieldset';
-import { Fieldset } from '../../Fieldset';
+import type { FieldsetProps } from '../Fieldset';
+import { Fieldset } from '../Fieldset';
 
 export type RadioGroupContextProps = {
   name?: string;
