+++ packages/react/src/components/form/Radio/useRadio.ts
@@ -4,8 +4,8 @@ import { useContext } from 'react';
 import type { FormField } from '../useFormField';
 import { useFormField } from '../useFormField';
 
-import { RadioGroupContext } from './Group';
 import type { RadioProps } from './Radio';
+import { RadioGroupContext } from './RadioGroup';
 
 type UseRadio = (props: RadioProps) => FormField & {
   inputProps?: Pick<
